+++ lib/ch/connection.ex
@@ -190,20 +190,46 @@ defmodule Ch.Connection do
   end
 
   @impl true
-  def handle_execute(%Query{statement: statement} = query, params, opts, conn) do
+  def handle_execute(%Query{statement: statement} = query, {:stream, params}, opts, conn) do
     {query_params, extra_headers} = params
 
     path = path(conn, query_params, opts)
     headers = headers(conn, extra_headers, opts)
 
-    result =
-      if is_list(statement) or is_binary(statement) do
-        request(conn, "POST", path, headers, statement, opts)
-      else
-        request_chunked(conn, "POST", path, headers, statement, opts)
+    with {:ok, conn, ref} <- send_request(conn, "POST", path, headers, :stream) do
+      case HTTP.stream_request_body(conn, ref, statement) do
+        {:ok, conn} -> {:ok, query, ref, conn}
+        {:error, conn, reason} -> {:disconnect, reason, conn}
       end
+    end
+  end
 
-    with {:ok, conn, responses} <- result do
+  def handle_execute(%Query{} = query, {:stream, ref, body}, opts, conn) do
+    case HTTP.stream_request_body(conn, ref, body) do
+      {:ok, conn} ->
+        case body do
+          :eof ->
+            with {:ok, conn, responses} <- receive_full_response(conn, timeout(conn, opts)) do
+              {:ok, query, responses, conn}
+            end
+
+          _other ->
+            {:ok, query, ref, conn}
+        end
+
+      {:error, conn, reason} ->
+        {:disconnect, reason, conn}
+    end
+  end
+
+  def handle_execute(%Query{statement: statement} = query, params, opts, conn)
+      when is_list(statement) or is_binary(statement) do
+    {query_params, extra_headers} = params
+
+    path = path(conn, query_params, opts)
+    headers = headers(conn, extra_headers, opts)
+
+    with {:ok, conn, responses} <- request(conn, "POST", path, headers, statement, opts) do
       {:ok, query, responses, conn}
     end
   end
@@ -230,40 +256,6 @@ defmodule Ch.Connection do
   defp request(conn, method, path, headers, body, opts) do
     with {:ok, conn, _ref} <- send_request(conn, method, path, headers, body) do
       receive_full_response(conn, timeout(conn, opts))
-    end
-  end
-
-  @spec request_chunked(
-          conn,
-          method :: String.t(),
-          path :: String.t(),
-          Mint.Types.headers(),
-          body :: Enumerable.t(),
-          [Ch.query_option()]
-        ) ::
-          {:ok, conn, [response]}
-          | {:error, Error.t(), conn}
-          | {:disconnect, Mint.Types.error(), conn}
-  def request_chunked(conn, method, path, headers, stream, opts) do
-    with {:ok, conn, ref} <- send_request(conn, method, path, headers, :stream),
-         {:ok, conn} <- stream_body(conn, ref, stream),
-         do: receive_full_response(conn, timeout(conn, opts))
-  end
-
-  @spec stream_body(conn, Mint.Types.request_ref(), Enumerable.t()) ::
-          {:ok, conn} | {:disconnect, Mint.Types.error(), conn}
-  defp stream_body(conn, ref, stream) do
-    result =
-      stream
-      |> Stream.concat([:eof])
-      |> Enum.reduce_while({:ok, conn}, fn
-        chunk, {:ok, conn} -> {:cont, HTTP.stream_request_body(conn, ref, chunk)}
-        _chunk, {:error, _conn, _reason} = error -> {:halt, error}
-      end)
-
-    case result do
-      {:ok, _conn} = ok -> ok
-      {:error, conn, reason} -> {:disconnect, reason, conn}
     end
   end